# lib/api-zod/src/generated/types/calendarResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Taskify API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarDay } from './calendarDay';
import type { CalendarResponseStatusSummary } from './calendarResponseStatusSummary';
import type { Task } from './task';

export interface CalendarResponse {
  days: CalendarDay[];
  todayTasks: Task[];
  statusSummary: CalendarResponseStatusSummary;
  quote: string;
}