# lib/api-zod/src/generated/types/taskListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Taskify API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Task } from './task';

export interface TaskListResponse {
  tasks: Task[];
  total: number;
  page: number;
  totalPages: number;
}